Haven/law/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
python-multipart==0.0.6
jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 